% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{big_test_res}
\alias{big_test_res}
\title{Test summary result}
\format{
## `big_test_res` A tibble from posterior::summarise_draws(test_ddf)
}
\usage{
big_test_res
}
\description{
A big test summary result to check that fsummary gets the same numbers. No accompanying
ddf, see the generating script in data-raw/
}
\keyword{datasets}
