% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{na_ddf}
\alias{na_ddf}
\title{zero variance draws data frame}
\format{
## `na_ddf`
A draws_df of three parameters, V1, V2, and V3, all numeric with 100 iterations and four chains.
}
\usage{
na_ddf
}
\description{
A test draws data frame where some parameters have 0 variance, which should produce some NAs.
}
\keyword{datasets}
