% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{na_res}
\alias{na_res}
\title{zero variance summary result}
\format{
## `na_res`
A tibble from posterior::summarise_draws(test_ddf)
}
\usage{
na_res
}
\description{
A test summary result where some parameters have 0 variance, which should produce some NAs.
}
\keyword{datasets}
