% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_ddf}
\alias{test_ddf}
\title{Test draws data frame}
\format{
## `test_ddf`
A draws_df of three parameters, V1, V2, and V3, all numeric with 100 iterations and four chains.
}
\usage{
test_ddf
}
\description{
A test draws data frame.
}
\keyword{datasets}
