% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_res}
\alias{test_res}
\title{Test summary result}
\format{
## `test_res`
A tibble from posterior::summarise_draws(test_ddf)
}
\usage{
test_res
}
\description{
A test summary result to check that fsummary gets the same numbers.
}
\keyword{datasets}
