% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsummary.R
\name{fsummary}
\alias{fsummary}
\title{Fast summary function for cmdstanr draws}
\usage{
fsummary(ddf, conv_metrics = TRUE, verbose = FALSE)
}
\arguments{
\item{ddf}{a draws_df}

\item{conv_metrics}{logical indicating whether to compute convergence metrics (slower)}
}
\value{
a data.table of summary metrics
}
\description{
This function quickly computes the same thing as \link[posterior]{summarise_draws} with two minor differences:
\itemize{
  \item{it produces a \link[data.table]{data.table} instead of a \link[tibble]{tibble}}
  \item{there can be negligible floating point error way out past the decimal point}
}
}
\details{
\itemize{
\item{The input \code{ddf} can ONLY be a \link[posterior]{draws_df}. No arrays, lists, etc. Use \link[posterior]{as_draws_df} if needed.}
\item{Activate parallelization by calling \link[mirai]{daemons} with your desired
  number of cores before calling \code{fsummary}.}
}
}
\examples{
fsummary(test_ddf)

}
