% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsummary.R
\name{fsummary}
\alias{fsummary}
\title{Fast summary function for cmdstanr draws}
\usage{
fsummary(
  ddf,
  conv_metrics = TRUE,
  .cores = getOption("mc.cores", 1),
  verbose = FALSE
)
}
\arguments{
\item{ddf}{a draws_df}

\item{conv_metrics}{logical indicating whether to compute convergence metrics (slower)}

\item{.cores}{cores to use when computing mean/median}
}
\value{
a data.table of summary metrics
}
\description{
Fast summary function for cmdstanr draws
}
\details{
Currently .cores is only used by fmean and fmedian, so it likely won't help
much unless you have loads of parameters and/or extremely long chains.
}
